Panel Tasarımı Tamamlandı

+++ psql/kullanici_davranislari/kullanici_davranis_loglari.sql
@@ -6,6 +6,6 @@ CREATE TABLE IF NOT EXISTS sayfa_ziyaret_loglari (
     oturum_token    TEXT NOT NULL,
     giris_zamani    TIMESTAMP WITHOUT TIME ZONE DEFAULT CURRENT_TIMESTAMP,
     son_guncelleme  TIMESTAMP WITHOUT TIME ZONE DEFAULT CURRENT_TIMESTAMP,
-    sure_saniye     INTEGER DEFAULT 0,
+    sure_saniye     NUMERIC(12,3) DEFAULT 0,
     FOREIGN KEY (sicil) REFERENCES kullanicilar(sicil)
 );
